refactor(economy): phase 3 slice 28 FRR runner merge and validator presets (#3939)
Consolidate D5 AC audit groups into deal-matcher and FRR credits runners,
merge non-GP extraction part runners, and add validator catalog-default
presets for treasury-cap scenario rows.

Refs #3939
