+++ tool/check_economy_scenario_table_runner.dart
@@ -16,8 +16,7 @@ const _maxImperativeTestBodyLines = 15;
 const _allowlistedRelativePaths = <String>{
   'packages/colonizethis_economy/test/resource_extractor_test.dart',
   'packages/colonizethis_economy/test/economy/world_market/world_market_deal_matcher_test.dart',
-  'packages/colonizethis_economy/test/economy/world_market/first_right_of_refusal_issue_acceptance_criteria_d5_test.dart',
-  'packages/colonizethis_economy/test/economy/world_market/first_right_profit_test.dart',
+  'packages/colonizethis_economy/test/economy/world_market/first_right_credits_test.dart',
   'packages/colonizethis_economy/test/economy/world_market/world_market_treasury_test.dart',
   'packages/colonizethis_economy/test/economy/world_market/world_market_trade_order_validator_test.dart',
 };
